+++ src/main/java/org/example/repositories/LineRepository.java
@@ -3,5 +3,5 @@ package org.example.repositories;
 import org.example.models.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LineRepository extends JpaRepository<Line, Integer> {
+public interface LineRepository extends JpaRepository<Line, Long> {
 }
